fix: Adiciona atributo avatar somente no veterinario
Issue#28
