+++ src/main/java/org/springframework/samples/petclinic/model/Person.java
@@ -38,10 +38,6 @@ public class Person extends BaseEntity {
 	@NotEmpty
 	private String lastName;
 
-	@Lob
-	@Column(name = "avatar", length = Integer.MAX_VALUE, nullable = true)
-	private byte[] avatar;
-
 	public String getFirstName() {
 		return this.firstName;
 	}
@@ -56,14 +52,6 @@ public class Person extends BaseEntity {
 
 	public void setLastName(String lastName) {
 		this.lastName = lastName;
-	}
-
-	public byte[] getAvatar() {
-		return this.avatar;
-	}
-
-	public void setAvatar(byte[] avatar) {
-		this.avatar = avatar;
 	}
 
 }
